Make MeterCache concrete
Since it's not planned to be pluggable, make it concreted and check for nullity
